fix(): move story and add creevey test
