+++ packages/react-ui/components/Radio/__stories__/Radio.stories.tsx
@@ -1,24 +1,10 @@
 // TODO: Rewrite stories and enable rule (in process of functional refactoring).
 /* eslint-disable react/no-unstable-nested-components */
-import React, { FormEvent, useCallback, useRef, useState } from 'react';
-import { globalObject } from '@skbkontur/global-object';
+import React from 'react';
 
 import { Meta, Story } from '../../../typings/stories';
 import { Gapped } from '../../Gapped';
 import { Radio } from '../Radio';
-import { Input } from '../../Input';
-import { Autocomplete } from '../../Autocomplete';
-import { Button } from '../../Button';
-import { THEME_2022 } from '../../../lib/theming/themes/Theme2022';
-import { ThemeContext } from '../../../lib/theming/ThemeContext';
-import { CurrencyInput } from '../../CurrencyInput';
-import { Checkbox } from '../../Checkbox';
-import { DateInput } from '../../DateInput';
-import { DatePicker } from '../../DatePicker';
-import { FileUploader } from '../../FileUploader';
-import { FxInput } from '../../FxInput';
-import { Toggle } from '../../Toggle';
-import { Switcher } from '../../Switcher';
 
 export default {
   title: 'Radio',
@@ -53,98 +39,6 @@ RadioWithDifferentStates.parameters = {
     },
   },
 };
-
-export const RadioFocus = () => {
-  const [isDisabled, setIsDisabled] = useState(false);
-  const inputRef = useRef<Input>(null);
-  const onSubmit = (e: FormEvent) => {
-    e.preventDefault();
-    e.stopPropagation();
-    new Promise((resolve) => {
-      setIsDisabled(true);
-      setTimeout(resolve, 1000);
-    }).then(() => {
-      setIsDisabled(false);
-    });
-  };
-
-  const handleKeyDown = (e: KeyboardEvent) => {
-    if (e.key === 'Enter') {
-      inputRef.current?.blur();
-      new Promise((resolve) => {
-        setIsDisabled(true);
-        setTimeout(resolve, 1000);
-      }).then(() => {
-        setIsDisabled(false);
-      });
-    }
-  };
-  globalObject.document?.addEventListener('keydown', handleKeyDown);
-
-  const emptyFunction = useCallback(() => {
-    return;
-  }, []);
-
-  return (
-    <>
-      <ThemeContext.Provider value={THEME_2022}>
-        <h1 style={{ margin: '10px' }}>Форма</h1>
-        <Gapped gap={5} style={{ border: 'black solid 1px', margin: '10px', padding: '10px' }}>
-          <form onSubmit={onSubmit}>
-            <div>
-              <Input disabled={isDisabled} value={'Input'}></Input>
-            </div>
-            <br />
-            <div>
-              <Autocomplete onValueChange={emptyFunction} value="Autocomplete" disabled={isDisabled} />
-            </div>
-            <br />
-            <div>
-              <Switcher onValueChange={emptyFunction} disabled={isDisabled} items={['1', '2', '3']} />
-            </div>
-            <br />
-            <div>
-              <CurrencyInput onValueChange={emptyFunction} disabled={isDisabled} />
-            </div>
-            <br />
-            <div>
-              <Checkbox disabled={isDisabled} />
-            </div>
-            <br />
-            <div>
-              <DateInput disabled={isDisabled} />
-            </div>
-            <br />
-            <div>
-              <DatePicker onValueChange={emptyFunction} disabled={isDisabled} />
-            </div>
-            <br />
-            <div>
-              <FileUploader disabled={isDisabled} />
-            </div>
-            <br />
-            <div>
-              <FxInput onValueChange={emptyFunction} disabled={isDisabled} />
-            </div>
-            <br />
-            <div>
-              <Radio value={'Radio'} disabled={isDisabled} />
-            </div>
-            <br />
-            <div>
-              <Toggle disabled={isDisabled} />
-            </div>
-            <br />
-            <Button use="primary" type="submit">
-              Submit
-            </Button>
-          </form>
-        </Gapped>
-      </ThemeContext.Provider>
-    </>
-  );
-};
-RadioFocus.storyName = 'Radio focus';
 
 export const Playground = () => {
   class Comp extends React.Component {
